back/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
django-qrcode==0.3
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.95.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
loguru==0.7.0
multidict==6.0.4
Naked==0.1.32
openai==0.27.7
Pillow==9.5.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==1.10.8
pymongo==4.3.3
pyOpenSSL==23.1.1
pypng==0.20220715.0
python-alipay-sdk==3.2.0
PyYAML==6.0
qrcode==7.4.2
redis==4.5.5
requests==2.31.0
rsa==4.9
shellescape==3.8.1
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
wechatpayv3==1.2.39
yarl==1.9.2
